rename Atoms in the "create new" dropdown
We've received the following request:
```
Using the "Create new Atom" function in Workflow creates a video atom. This has led to confusion as people expect it to go to the Atom menu from where you can choose timeline/quick guide etc.

Could this be renamed "Create Video Atom" or better link to the Atom dashboard?
```

Ultimately, we probably want to link to https://atomworkshop.code.dev-gutools.co.uk/create and get rid of this form or allow more atoms to be created without leaving workflow. This is just a quick fix until we have time to properly investigate.