Remove "optimisation" to only issue queries once
This was originally in the code to attempt to ease pressure on the DB if lots of people were to subscribe to the same query. It's useless as is however since "email" is added to the query parameters (forfacilitate "assigned to me" searches) which personalises each query to the individual user anyway.