Fix sbt dependency graph workflow
At the moment, this workflow is [failing on
main](https://github.com/guardian/workflow-frontend/actions/runs/12430359668/job/34705650369)
because it doesn’t explicitly install sbt, and github has updated the
ubuntu-latest tag to the 24.04 runners which don’t include sbt by
default.

To fix this, this commit updates the workflow by copying the version
from https://github.com/guardian/fastly-cache-purger/pull/88, which PR
was produced by a newer version of the gu-dependency-graph-integrator
bot and includes an explicit installation of sbt (and updates to some
action versions).