Update actions-riff-raff to v4
actions-riff-raff v4 comments on the PR with some useful links for
deploying, so I’d like to update to it. I’ve followed the changes
described in [the
readme](https://github.com/guardian/actions-riff-raff/blob/6ddd4acd261c4265995908a68f0592ad1604b56e/README.md#L267-L286)
for migrating from v3 to v4, and copied the necessary changes for v3 as
well (to do with commenting on the PR).