Add type guard for value expected to be a string
[A recent error in
Sentry](https://the-guardian.sentry.io/issues/6362818459/?alert_rule_id=56564&alert_timestamp=1741252090194&alert_type=email&notification_uuid=58661c31-1ebe-47a1-927e-d679b91117f9&project=33134&referrer=alert_email)
appears to indicate that sectionsString here is truthy but not a string.
Since the code seems to expect it to be a string, this commit adds a
type guard to make sure we only use the value if it is a string.

It seems this value comes from the request URL, though I’m not 100% sure
how. I’m also not sure how it came to be a non-string value.