Update sbt version to 1.11.4
This commit updates the project’s sbt version to the latest version,
[1.11.4](https://github.com/sbt/sbt/releases/tag/v1.11.4).

This project is using version 1.9.6 of sbt, which [is now nearly two
years old](https://github.com/sbt/sbt/releases/v1.9.6). Using an old
version of sbt causes breakage of the `console` task on newer versions
of scala 2.13, and it’s probably good practice to be on a recent version
of sbt anyway.